+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased](https://github.com/kabamgamer/dd2-strategy-hub/tree/HEAD)
+
+* Fixed a bug in the UI causing collapse to snap instead of transition
+
+[Full Changelog](https://github.com/kabamgamer/dd2-strategy-hub/compare/v1.1.8...HEAD)
+
 ## [v1.1.8](https://github.com/kabamgamer/dd2-strategy-hub/tree/v1.1.8) (2023-08-29)
 
 * Hide shards/mods if they are incompatible with the defense